Implemented Turning on all 5 levels
Except for Level 1 which does not turn due to being an intro to the game itself.
New bug found, dying once to collision and only collision, turning is calculated with what seems to be an average of the turning of all the turners in the level and run for every reset from then on, soft locking you in that level due to not having a pause menu or a quit button. Will look into it as well. Forcing the triggers to require the player does not fix the bug.